baltic_layer_reader fixed indexes
Fixed the issue with fill blocks not finding correct locations due to index issues resulting in np.nan filled blocks;
to do this , I adjusted the get_array_from_tiff and get_value_from_array_functions. Now it rebuilds an index for the baltic sea maps and is able to extract the value of the array.